requirements.txt
apache-beam[gcp]==2.71.0
google-cloud-bigquery==3.17.2
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.42.1
pandas==2.1.4